+++ frontend/src/app/worker/selection.cljs
@@ -54,6 +54,7 @@
           match-criteria?
           (fn [shape]
             (and (not (:hidden shape))
+                 (not (:blocked shape))
                  (or (not frame-id) (= frame-id (:frame-id shape)))
                  (case (:type shape)
                    :frame   include-frames?